requirements.txt
pandas
openai==0.28
python-dotenv
streamlit